new llc reading is basically working
