+++ xmitgcm/llc_utils.py
@@ -53,7 +53,7 @@ def _read_2d_face(fname, nface, nlev, nx, dtype='>f8', memmap=True):
     return data
 
 # manually construct dask graph
-def read_3d_llc_array(fname, nz, nx, dtype='>f8', memmap=True):
+def read_3d_llc_data(fname, nz, nx, dtype='>f8', memmap=True):
     """Read a three-dimensional LLC file using a custom dask graph.
 
     PARAMETERS
@@ -95,3 +95,58 @@ def read_3d_llc_array(fname, nz, nx, dtype='>f8', memmap=True):
         data = data[0]
 
     return data
+
+
+# a deprecated function that I can't bear to delete because it was painful to
+# write
+def _reshape_llc_data(data, jdim):
+    """Fix the weird problem with llc data array order."""
+    # Can we do this without copying any data?
+    # If not, we need to go upstream and implement this at the MDS level
+    # Or can we fudge it with dask?
+    # this is all very specific to the llc file output
+    # would be nice to generalize more, but how?
+    nside = data.shape[jdim] / LLC_NUM_FACES
+    # how the LLC data is laid out along the j dimension
+    strides = ((0,3), (3,6), (6,7), (7,10), (10,13))
+    # whether to reshape each face
+    reshape = (False, False, False, True, True)
+    # this will slice the data into 5 facets
+    slices = [jdim * (slice(None),) + (slice(nside*st[0], nside*st[1]),)
+              for st in strides]
+    facet_arrays = [data[sl] for sl in slices]
+    face_arrays = []
+    for ar, rs, st in zip(facet_arrays, reshape, strides):
+        nfaces_in_facet = st[1] - st[0]
+        shape = list(ar.shape)
+        if rs:
+            # we assume the other horizontal dimension is immediately after jdim
+            shape[jdim] = ar.shape[jdim+1]
+            shape[jdim+1] = ar.shape[jdim]
+        # insert a length-1 dimension along which to concatenate
+        shape.insert(jdim, 1)
+        # this modify the array shape in place, with no copies allowed
+        # but it doesn't work with dask arrays
+        # ar.shape = shape
+        ar = ar.reshape(shape)
+        # now ar is propery shaped, but we still need to slice it into faces
+        face_slice_dim = jdim + 1 + rs
+        for n in range(nfaces_in_facet):
+            face_slice = (face_slice_dim * (slice(None),) +
+                          (slice(nside*n, nside*(n+1)),))
+            data_face = ar[face_slice]
+            face_arrays.append(data_face)
+
+    # We can't concatenate using numpy (hcat etc.) because it makes a copy,
+    # presumably loading the memmaps into memory.
+    # Using dask gets around this.
+    # But what if we want different chunks, or already chunked the data
+    # upstream? Doesn't seem like this is ideal
+    # TODO: Refactor handling of dask arrays and chunking
+    #return np.concatenate(face_arrays, axis=jdim)
+    # the dask version doesn't work because of this:
+    # https://github.com/dask/dask/issues/1645
+    face_arrays_dask = [da.from_array(fa, chunks=fa.shape)
+                        for fa in face_arrays]
+    concat = da.concatenate(face_arrays_dask, axis=jdim)
+    return concat
