+++ xmitgcm/test/test_mds_store.py
@@ -298,7 +298,7 @@ def test_read_mds_no_meta(all_mds_datadirs):
 def test_read_raw_data_llc(llc_mds_datadirs):
     dirname, expected = llc_mds_datadirs
 
-    from xmitgcm.llc_utils import read_3d_llc_array
+    from xmitgcm.llc_utils import read_3d_llc_data
 
     shape = expected['shape']
     nz, nface, ny, nx = shape
@@ -308,12 +308,12 @@ def test_read_raw_data_llc(llc_mds_datadirs):
     # if we use memmap=True, we open too many files
 
     fname = os.path.join(dirname, 'T.%010d.data' % expected['test_iternum'])
-    data = read_3d_llc_array(fname, nz, nx, dtype=dtype, memmap=False)
+    data = read_3d_llc_data(fname, nz, nx, dtype=dtype, memmap=False)
     assert data.shape == shape
     assert data.compute().shape == shape
 
     fname = os.path.join(dirname, 'XC.data')
-    data = read_3d_llc_array(fname, 1, nx, dtype=dtype, memmap=False)
+    data = read_3d_llc_data(fname, 1, nx, dtype=dtype, memmap=False)
     # make sure the first dimension is squeezed off
     assert data.shape == shape[1:]
     assert data.compute().shape == shape[1:]
