+++ xmitgcm/utils.py
@@ -12,6 +12,7 @@ from functools import reduce
 from dask import delayed
 import dask.array as dsa
 
+from .llc_utils import read_3d_llc_data
 
 def parse_meta_file(fname):
     """Get the metadata as a dict out of the mitGCM mds .meta file."""
@@ -52,7 +53,7 @@ def parse_meta_file(fname):
 
 
 def read_mds(fname, iternum=None, use_mmap=True, force_dict=True, endian='>',
-             shape=None, dtype=None, dask_delayed=True):
+             shape=None, dtype=None, dask_delayed=True, llc=False):
     """Read an MITgcm .meta / .data file pair"""
 
     if iternum is None:
@@ -80,6 +81,8 @@ def read_mds(fname, iternum=None, use_mmap=True, force_dict=True, endian='>',
         else:
             name = meta['basename']
     except IOError as e:
+        # we can recover from not having a .meta file if dtype and shape have
+        # been specified already
         if (shape is None) or (dtype is None):
             raise e
         else:
@@ -88,7 +91,20 @@ def read_mds(fname, iternum=None, use_mmap=True, force_dict=True, endian='>',
             shape.insert(0, nrecs)
             name = os.path.basename(fname)
 
-    if dask_delayed:
+    # this will exclude vertical profile files
+    if llc and shape[-1]>1:
+        # remeberer that the first dim is nrec
+        if nrecs>1:
+            raise ValueError("For now, can't handle nrecs>1 with llc==True.")
+        if len(shape)==4:
+            _, nz, ny, nx = shape
+        else:
+            _, ny, nx = shape
+            nz = 1
+        d = read_3d_llc_data(datafile, nz, nx, dtype=dtype, memmap=False)
+        # add record dimension
+        d = d[None]
+    elif dask_delayed:
         d = dsa.from_delayed(
               delayed(read_raw_data)(datafile, dtype, shape, use_mmap=use_mmap),
               shape, dtype
